+++ demos/manual/src/main/java/is/codion/manual/common/demo/ApplicationPanel.java
@@ -446,12 +446,12 @@ public final class ApplicationPanel extends JPanel {
 
 		@Override
 		public Integer value(Item<Integer> item) {
-			return item.value();
+			return item.getOrThrow();
 		}
 
 		@Override
 		public Predicate<Item<Integer>> predicate(Integer value) {
-			return item -> Objects.equals(item.value(), value);
+			return item -> Objects.equals(item.get(), value);
 		}
 	}