+++ demos/manual/src/main/java/is/codion/manual/keybinding/KeyBindingModel.java
@@ -133,7 +133,7 @@ final class KeyBindingModel {
 		@Override
 		public Collection<String> get() {
 			return lookAndFeelModel.selection().item().optional()
-							.map(Item::value)
+							.map(Item::get)
 							.map(LookAndFeelEnabler::lookAndFeel)
 							.map(LookAndFeel::getDefaults)
 							.map(Hashtable::keySet)