+++ framework/domain-test/src/main/java/is/codion/framework/domain/test/DefaultEntityFactory.java
@@ -264,7 +264,7 @@ public class DefaultEntityFactory implements EntityFactory {
 		List<Item<T>> items = attributeDefinition.items();
 		Item<T> item = items.get(RANDOM.nextInt(items.size()));
 
-		return item.value();
+		return item.get();
 	}
 
 	private static int randomInteger(AttributeDefinition<?> attributeDefinition) {