+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/AbstractAttributeDefinition.java
@@ -112,7 +112,7 @@ abstract class AbstractAttributeDefinition<T> implements AttributeDefinition<T>,
 		this.dateTimeFormatter = builder.dateTimeFormatter;
 		this.items = builder.items;
 		this.itemMap = items == null ? null : items.stream()
-						.collect(toMap(Item::value, Function.identity()));
+						.collect(toMap(Item::get, Function.identity()));
 	}
 
 	@Override
@@ -407,7 +407,7 @@ abstract class AbstractAttributeDefinition<T> implements AttributeDefinition<T>,
 
 		private ItemComparator(List<Item<T>> items) {
 			this.captions = items.stream()
-							.collect(toMap(Item::value, Item::caption));
+							.collect(toMap(Item::get, Item::caption));
 		}
 
 		@Override