+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTable.java
@@ -915,11 +915,11 @@ public final class FilterTable<R, C> extends JTable {
 		List<ToggleControl> controls = new ArrayList<>();
 		State.Group group = State.group();
 		for (Item<Integer> resizeMode : AUTO_RESIZE_MODES) {
-			State state = State.state(resizeMode.value().equals(getAutoResizeMode()));
+			State state = State.state(resizeMode.getOrThrow().equals(getAutoResizeMode()));
 			group.add(state);
 			state.addConsumer(enabled -> {
 				if (enabled) {
-					setAutoResizeMode(resizeMode.value());
+					setAutoResizeMode(resizeMode.getOrThrow());
 				}
 			});
 			controls.add(Control.builder()